Refactor getClanMembers function and delete fullClan type
